azurerm/internal/services/applicationinsights/parse/application_insights_api_key_test.go
package parse

import (
	"testing"
)

func TestApplicationInsightsApiKeyID(t *testing.T) {
	testData := []struct {
		Name     string
		Input    string
		Expected *ApplicationInsightsApiKeyId
	}{
		{
			Name:     "Empty",
			Input:    "",
			Expected: nil,
		},
		{
			Name:     "No Resource Groups Segment",
			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000",
			Expected: nil,
		},
		{
			Name:     "No Resource Groups Value",
			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/",
			Expected: nil,
		},
		{
			Name:     "Resource Group ID",
			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/",
			Expected: nil,
		},
		{Name: "Missing Components Value",
			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/resGroup1/providers/Microsoft.Insights/components/",
			Expected: nil,
		},
		{
			Name:     "Application Insights ID",
			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/resGroup1/providers/Microsoft.Insights/components/component1",
			Expected: nil,
		},
		{
			Name:     "Missing API Key Value",
			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/resGroup1/providers/Microsoft.Insights/components/component1/apikeys/",
			Expected: nil,
		},
		{
			Name:  "App Insights API Key ID",
			Input: "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/resGroup1/providers/Microsoft.Insights/components/component1/apikeys/key1",
			Expected: &ApplicationInsightsApiKeyId{
				ResourceGroup: "resGroup1",
				Name:          "component1",
				KeyID:         "key1",
			},
		},
		{
			Name:     "Wrong Casing",
			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/resGroup1/providers/Microsoft.Insights/components/component1/Apikeys/key1",
			Expected: nil,
		},
	}

	for _, v := range testData {
		t.Logf("[DEBUG] Testing %q", v.Name)

		actual, err := ApplicationInsightsApiKeyID(v.Input)
		if err != nil {
			if v.Expected == nil {
				continue
			}

			t.Fatalf("Expected a value but got an error: %s", err)
		}

		if actual.ResourceGroup != v.Expected.ResourceGroup {
			t.Fatalf("Expected %q but got %q for Resource Group", v.Expected.ResourceGroup, actual.ResourceGroup)
		}
		if actual.Name != v.Expected.Name {
			t.Fatalf("Expected %q but got %q for Name", v.Expected.Name, actual.Name)
		}
		if actual.KeyID != v.Expected.KeyID {
			t.Fatalf("Expected %q but got %q for API Key ID", v.Expected.KeyID, actual.KeyID)
		}

	}
}
